CS50x/Backend/vector-service/requirements.txt
fastapi
uvicorn
numpy
faiss-cpu